refactor(training): added base class for multidataset and multidomain
