+++ training/src/anemoi/training/data/dataset.py
@@ -1,0 +1,265 @@
+# (C) Copyright 2024 Anemoi contributors.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
+import datetime
+import logging
+from abc import ABC
+from abc import abstractmethod
+from functools import cached_property
+from typing import TYPE_CHECKING
+
+# Move third-party import inside this block
+if TYPE_CHECKING:
+    import numpy as np
+
+from rich.console import Console
+from rich.tree import Tree
+from torch.utils.data import IterableDataset
+
+from anemoi.models.distributed.balanced_partition import get_balanced_partition_sizes
+from anemoi.models.distributed.balanced_partition import get_partition_range
+from anemoi.training.data.data_reader import BaseAnemoiReader
+
+LOGGER = logging.getLogger(__name__)
+
+
+class AnemoiDataset(IterableDataset, ABC):
+    """Base Anemoi Datasets torch dataset class."""
+
+    def __init__(
+        self,
+        data_readers: dict[str, BaseAnemoiReader],
+        shuffle: bool = True,
+        label: str = "multi",
+    ) -> None:
+        """Initialize multi-dataset with synchronized data readers.
+
+        Parameters
+        ----------
+        data_readers : dict[str, BaseAnemoiReader]
+            Dictionary mapping dataset names to their data_readers
+            Format: {"dataset_a": data_reader_a, "dataset_b": data_reader_b, ...}
+        shuffle : bool, optional
+            Shuffle batches, by default True
+        label : str, optional
+            label for the dataset, by default "multi"
+        """
+        self.data_readers = data_readers
+        self.label = label
+        self.shuffle = shuffle
+        self.dataset_names = list(data_readers.keys())
+        self._lazy_init_model_and_reader_group_info()
+
+    def _lazy_init_model_and_reader_group_info(self) -> None:
+        """Lazy initialize model and reader group info."""
+        # lazy init model and reader group info, will be set by the DDPGroupStrategy:
+        self.model_comm_group_rank = 0
+        self.model_comm_num_groups = 1
+        self.model_comm_group_id = 0
+        self.global_rank = 0
+
+        self.reader_group_rank = 0
+        self.reader_group_size = 1
+
+        self.sample_comm_num_groups = 1  # groups that work on the same sample / batch
+        self.sample_comm_group_id = 0
+
+        self.ens_comm_group_rank = 0
+        self.ens_comm_num_groups = 1
+        self.ens_comm_group_id = 0
+
+        self.shard_shapes = None
+
+        # additional state vars (lazy init)
+        self.n_samples_per_worker = 0
+        self.chunk_index_range: np.ndarray | None = None
+
+    def _collect(self, attr_name: str) -> dict:
+        """Helper method to collect attributes from all data readers."""
+        return {name: getattr(dataset, attr_name) for name, dataset in self.data_readers.items()}
+
+    @cached_property
+    def statistics(self) -> dict[str, dict]:
+        """Return combined statistics from all data readers."""
+        return self._collect("statistics")
+
+    @cached_property
+    def metadata(self) -> dict[str, dict]:
+        """Return combined metadata from all data readers."""
+        return self._collect("metadata")
+
+    @cached_property
+    def supporting_arrays(self) -> dict[str, dict]:
+        """Return combined supporting arrays from all data readers."""
+        return self._collect("supporting_arrays")
+
+    @cached_property
+    def variables(self) -> dict[str, list[str]]:
+        """Return combined variables from all data readers."""
+        return self._collect("variables")
+
+    @property
+    def data(self) -> dict:
+        """Return data from all data readers as dictionary."""
+        return self._collect("data")
+
+    @cached_property
+    def name_to_index(self) -> dict[str, dict]:
+        """Return combined name_to_index mapping from all data readers."""
+        return self._collect("name_to_index")
+
+    @cached_property
+    def resolution(self) -> dict[str, str]:
+        """Return combined resolution from all data readers."""
+        return self._collect("resolution")
+
+    @cached_property
+    def frequency(self) -> datetime.timedelta:
+        """Return combined frequency from all data readers."""
+        freqs = self._collect("frequency")
+        freq_ref = None
+        for name, freq in freqs.items():
+            if freq_ref is None:
+                freq_ref = freq
+            assert freq == freq_ref, f"Data reader '{name}' has different frequency than other data readers"
+        return freq_ref
+
+    def set_comm_group_info(
+        self,
+        global_rank: int,
+        model_comm_group_id: int,
+        model_comm_group_rank: int,
+        model_comm_num_groups: int,
+        reader_group_rank: int,
+        reader_group_size: int,
+        shard_shapes: dict[str, list[int]],
+    ) -> None:
+        """Set model and reader communication group information (called by DDPGroupStrategy).
+
+        Parameters
+        ----------
+        global_rank : int
+            Global rank
+        model_comm_group_id : int
+            Model communication group ID
+        model_comm_group_rank : int
+            Model communication group rank
+        model_comm_num_groups : int
+            Number of model communication groups
+        reader_group_rank : int
+            Reader group rank
+        reader_group_size : int
+            Reader group size
+        shard_shapes : dict[str, list[int]]
+            Shard shapes for all data readers
+        """
+        self.global_rank = global_rank
+        self.model_comm_group_id = model_comm_group_id
+        self.model_comm_group_rank = model_comm_group_rank
+        self.model_comm_num_groups = model_comm_num_groups
+        self.reader_group_rank = reader_group_rank
+        self.reader_group_size = reader_group_size
+
+        self.sample_comm_group_id = model_comm_group_id
+        self.sample_comm_num_groups = model_comm_num_groups
+
+        self.shard_shapes = shard_shapes
+
+        assert self.reader_group_size >= 1, f"reader_group_size(={self.reader_group_size}) must be positive"
+
+        LOGGER.info(
+            "NativeGridDataset.set_group_info(): global_rank %d, model_comm_group_id %d, "
+            "model_comm_group_rank %d, model_comm_num_groups %d, reader_group_rank %d, "
+            "sample_comm_group_id %d, sample_comm_num_groups %d",
+            global_rank,
+            model_comm_group_id,
+            model_comm_group_rank,
+            model_comm_num_groups,
+            reader_group_rank,
+            self.sample_comm_group_id,
+            self.sample_comm_num_groups,
+        )
+
+    def set_ens_comm_group_info(
+        self,
+        ens_comm_group_id: int,
+        ens_comm_group_rank: int,
+        ens_comm_num_groups: int,
+    ) -> None:
+        """Set ensemble communication group information (called by DDPGroupStrategy).
+
+        Parameters
+        ----------
+        ens_comm_group_id : int
+            Ensemble communication group ID
+        ens_comm_group_rank : int
+            Ensemble communication group rank
+        ens_comm_num_groups : int
+            Number of ensemble communication groups
+        """
+        self.ens_comm_group_id = ens_comm_group_id
+        self.ens_comm_group_rank = ens_comm_group_rank
+        self.ens_comm_num_groups = ens_comm_num_groups
+
+        self.sample_comm_group_id = ens_comm_group_id
+        self.sample_comm_num_groups = ens_comm_num_groups
+
+        LOGGER.info(
+            "NativeGridDataset.set_ens_comm_group_info(): global_rank %d, ens_comm_group_id %d, "
+            "ens_comm_group_rank %d, ens_comm_num_groups %d, reader_group_rank %d, "
+            "sample_comm_group_id %d, sample_comm_num_groups %d",
+            self.global_rank,
+            ens_comm_group_id,
+            ens_comm_group_rank,
+            ens_comm_num_groups,
+            self.reader_group_rank,
+            self.sample_comm_group_id,
+            self.sample_comm_num_groups,
+        )
+
+    @abstractmethod
+    def per_worker_init(self, n_workers: int, worker_id: int) -> None:
+        """Initialize all data readers for this worker. To be overwritten by subclasses."""
+
+    @cached_property
+    def shard_shapes(self) -> dict[str, list]:
+        """Return shard shapes for all data readers."""
+        shard_shapes = {}
+        for name, dataset in self.data_readers.items():
+            shard_shapes[name] = get_balanced_partition_sizes(dataset.grid_size, self.reader_group_size)
+        return shard_shapes
+
+    def get_shard_slice(self, dataset_name: str, reader_group_rank: int) -> slice:
+        """Get the grid shard slice according to the reader rank."""
+        start, end = get_partition_range(
+            partition_sizes=self.shard_shapes[dataset_name],
+            partition_id=reader_group_rank,
+        )
+        return slice(start, end)
+
+    @abstractmethod
+    def get_sample(self, index: int) -> None:
+        """Get a sample from data readers at the specified index. To be overwritten by subclasses."""
+
+    @abstractmethod
+    def __iter__(self) -> None:
+        """Return an iterator over the dataset(s). To be overwritten by subclasses."""
+
+    def __repr__(self) -> str:
+        console = Console(record=True, width=120)
+        with console.capture() as capture:
+            console.print(self.tree())
+        return capture.get()
+
+    def tree(self) -> Tree:
+        tree = Tree(f"{self.__class__.__name__}")
+        for name, dataset in self.data_readers.items():
+            subtree = dataset.tree(prefix=name)
+            tree.add(subtree)
+        return tree
